Initial Implementation for Systemd  (#2)
* added initial implementation for systemd

This module is currently experimental as it relies on dbus communication with systemd
This module will likely fail to build for MacOS as dbus is not available for Macs.
We should start separating modules/crates based on the OS build

* restricted builds to linux for the Systemd module.
Added skeletop for systemd module with the needed interfaces and properties for the runner
+ Some fix for CirrusCI

* fix for the CI pre_build script for systemd dbus package

* CI should be fixed now

* fixed package name for the CI

* yet another fix for the CI

* Implemented Start and Enable for services.
Current implementations does not have any error handling and will likely panic on any fault.
proper error handling is needed for this module to be considered stable.

* finished implementing disabled and restarted

* implemented custom error for the runner trait.
This should ensure that we utilize the question mark operator to help
with error handling.
An implementation for user's decision on error handling is needed. The
current implementations for various modules are a mix between ignore and
exit on error.

* implemented stopped for systemd
+ Added some changes for debugging with the ? operator. It might be best
  to change the implementation to implement From<> for the error enum.
Other options such as SNAFU might be helpful.

* moved shell tests to their own folders

* added option for service enablement.
This will allow the user to choose from whether they want to start the
unit file without enabling the service entirely.
Additionally, the unit getter was changed to LoadUnit due to the
limitation of GetUnit method when it comes to a disable service.